refactor venue filtering logic and extract constants for better maintainability
