+++ src/components/filters/FilterSort.jsx
@@ -8,47 +8,9 @@ import {
   MenuItems,
 } from "@headlessui/react";
 import { ChevronDownIcon, FunnelIcon } from "@heroicons/react/20/solid";
-
-const filtersData = {
-  price: [
-    { value: "0-500", label: "$0 - $500" },
-    { value: "500-1000", label: "$500 - $1000" },
-    { value: "1000-2000", label: "$1000 - $2000" },
-    { value: "2000+", label: "$2000+" },
-  ],
-  maxGuests: [
-    { value: "1-2", label: "1-2 Guests" },
-    { value: "3-4", label: "2-4 Guests" },
-    { value: "5-6", label: "5-6 Guests" },
-    { value: "7+", label: "7+ Guests" },
-  ],
-  rating: [
-    { value: "1", label: "1 star" },
-    { value: "2", label: "2 stars" },
-    { value: "3", label: "3 stars" },
-    { value: "4", label: "4 stars" },
-    { value: "5", label: "5 stars" },
-  ],
-  amenities: [
-    { value: "wifi", label: "WiFi" },
-    { value: "parking", label: "Parking" },
-    { value: "breakfast", label: "Breakfast" },
-    { value: "pets", label: "Pets" },
-  ],
-};
-
-const filterLabels = {
-  price: "Price",
-  maxGuests: "Capacity", // Change the label for maxGuests
-  rating: "Rating",
-  amenities: "Amenities",
-};
-
-const sortOptions = [
-  { name: "Most Popular", href: "#", current: true },
-  { name: "Best Rating", href: "#", current: false },
-  { name: "Newest", href: "#", current: false },
-];
+import { filtersData } from "../../constants/filtersData";
+import { filterLabels } from "../../constants/filterLabels";
+import { sortOptions } from "../../constants/sortOptions";
 
 function classNames(...classes) {
   return classes.filter(Boolean).join(" ");
