+++ src/constants/filterLabels.js
@@ -1,0 +1,6 @@
+export const filterLabels = {
+  price: "Price",
+  maxGuests: "Capacity", // Change the label for maxGuests
+  rating: "Rating",
+  amenities: "Amenities",
+};
