+++ src/constants/sortOptions.js
@@ -1,0 +1,5 @@
+export const sortOptions = [
+  { name: "Most Popular", href: "#", current: true },
+  { name: "Best Rating", href: "#", current: false },
+  { name: "Newest", href: "#", current: false },
+];
