+++ src/pages/Venues.jsx
@@ -1,10 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 import useFetch from "../hooks/useFetch";
 import { API_BASE_URL, API_VENUES_URL } from "../constants/apiUrls";
 import FilterSort from "../components/filters/FilterSort";
 import VenueSkeleton from "../components/loaders/VenueSkeleton";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faStar } from "@fortawesome/free-solid-svg-icons";
+import useVenueFilters from "../hooks/useVenueFilters";
 
 export default function Venues() {
   const {
@@ -12,73 +13,7 @@ export default function Venues() {
     loading,
     error,
   } = useFetch(`${API_BASE_URL}${API_VENUES_URL}`);
-  const [filters, setFilters] = useState({
-    price: [],
-    maxGuests: [],
-    rating: [],
-    amenities: [],
-  });
-
-  const clearFilters = () => {
-    setFilters({
-      price: [],
-      maxGuests: [],
-      rating: [],
-      amenities: [],
-    });
-  };
-
-  const applyFilters = (venues) => {
-    if (!venues) return [];
-    console.log("Applying filters:", filters);
-    return venues.data.filter((venue) => {
-      // Apply price filter
-      if (filters.price.length > 0) {
-        const price = parseFloat(venue.price);
-        if (
-          !filters.price.some((range) => {
-            if (range === "2000+") {
-              return price >= 2000;
-            }
-            const [min, max] = range.split("-").map(Number);
-            return price >= min && (max ? price <= max : true);
-          })
-        ) {
-          return false;
-        }
-      }
-      // Apply maxGuests filter
-      if (filters.maxGuests.length > 0) {
-        const guests = parseInt(venue.maxGuests);
-        if (
-          !filters.maxGuests.some((range) => {
-            if (range === "7+") {
-              return guests >= 7;
-            }
-            const [min, max] = range.split("-").map(Number);
-            return guests >= min && (max ? guests <= max : true);
-          })
-        ) {
-          return false;
-        }
-      }
-      // Apply rating filter
-      if (
-        filters.rating.length > 0 &&
-        !filters.rating.includes(venue.rating.toString())
-      ) {
-        return false;
-      }
-      // Apply amenities filter
-      if (
-        filters.amenities.length > 0 &&
-        !filters.amenities.every((amenity) => venue.meta[amenity])
-      ) {
-        return false;
-      }
-      return true;
-    });
-  };
+  const { filters, setFilters, clearFilters, applyFilters } = useVenueFilters();
 
   if (loading) {
     return <VenueSkeleton />;
@@ -88,7 +23,6 @@ export default function Venues() {
     return <p>{error.message}</p>;
   }
 
-  console.log("Venues data:", venues);
   const filteredVenues = applyFilters(venues);
 
   return (
